ingestion to its own module
